% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{as_ISO8601_parsed}
\alias{as_ISO8601_parsed}
\title{Create an ISO8601_parsed object}
\usage{
as_ISO8601_parsed(x)
}
\arguments{
\item{x}{a list from parse_ISO8601}
}
\description{
Create an ISO8601_parsed object
}
