% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{as_ISO8660_parsed}
\alias{as_ISO8660_parsed}
\title{Create an ISO8660_parsed object}
\usage{
as_ISO8660_parsed(x)
}
\arguments{
\item{x}{a list from parse_ISO8660}
}
\description{
Create an ISO8660_parsed object
}
