% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{build_parser}
\alias{build_parser}
\title{Create an ISO 8601 parser with options controlling which optional parts are included in the parser.}
\usage{
build_parser(
  allow_leap_second = FALSE,
  allow_before_1583 = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{allow_leap_second}{Should leap seconds be allowed (61 seconds in a minute)?}

\item{allow_before_1583}{Should years before 1583 be allowed?}

\item{verbose}{Show verbose messages while parsing?}
}
\description{
Create an ISO 8601 parser with options controlling which optional parts are included in the parser.
}
