% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISO8601.parser-package.R
\docType{package}
\name{ISO8601.parser-package}
\alias{ISO8601.parser}
\alias{ISO8601.parser-package}
\title{ISO8601.parser: Parses ISO 8601 date/time strings}
\description{
Parse ISO 8601 date/time strings and extract the relevant parts.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
