% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_ISO8660}
\alias{parse_ISO8660}
\title{Parse a single ISO8601 string}
\usage{
parse_ISO8660(x, parser)
}
\arguments{
\item{x}{The string to parse}

\item{parser}{The parser to use for parsing (as created by build_parser()).
If not given, the default parser will be used.}
}
\value{
A named list with the parsed parts
}
\description{
Parse a single ISO8601 string
}
